if not master

+++ Jenkinsfile
@@ -216,17 +216,8 @@ pipeline {
 
   post('Remove VM') {
     cleanup {
-      stages{
-        stage ('Cleanup'){
-          when {
-            not{
-             branch "master"
-            }
-          }
-          stage('Cleanup Branch') { 
-            sh 'docker-machine rm -y CEGA-${ID} LEGA-public-${ID} LEGA-private-${ID}'
-          }
-        }
+      if (env.BRANCH_NAME != "master") {
+        sh 'docker-machine rm -y CEGA-${ID} LEGA-public-${ID} LEGA-private-${ID}'
       }
     }
   }
